+++ .eslintignore
@@ -1,3 +1,4 @@
-/src/pages
+/src/options
+/src/popup
 /scripts/
 /build/
