+++ src/content_scripts/core/index.ts
@@ -1,0 +1,3 @@
+import onShortcutKeyUp from './keyboardShortcuts'
+
+document.addEventListener('keyup', onShortcutKeyUp)
