+++ src/content_scripts/utils/formatRuntime.ts
