+++ src/manifest.json
@@ -26,7 +26,7 @@
   "permissions": ["webNavigation", "storage"],
   "options_ui": {
     "open_in_tab": true,
-    "page": "/pages/options/options.html"
+    "page": "/options/options.html"
   },
   "background": {
     "service_worker": "/background.js"
@@ -36,19 +36,19 @@
       "matches": ["<all_urls>"],
       "all_frames": true,
       "run_at": "document_start",
-      "css": ["styles.css"],
+      "css": ["css/styles.css"],
       "js": ["content.js"]
     }
   ],
   "action": {
-    "default_popup": "pages/popup/popup.html",
+    "default_popup": "popup/popup.html",
     "default_title": "Toppings: Preferences",
     "default_icon": "assets/icons/icon128.png"
   },
   "web_accessible_resources": [
     {
       "matches":  ["<all_urls>"],
-      "resources": ["assets/icons/icon128.png", "styles.css", "pages/options/assets/icons8-settings-64.png"]
+      "resources": ["assets/icons/icon128.png", "css/styles.css" ,"options/assets/icons8-settings-64.png"]
     }
   ]
 }
