+++ src/options/options.html
@@ -9,7 +9,7 @@
 			href="./assets/icons8-settings-64.png"
 			type="image/x-icon"
 		/>
-		<link rel="stylesheet" href="../../styles.css" />
+		<link rel="stylesheet" href="vendor.css" />
 		<link rel="stylesheet" href="options.css" />
 		<link
 			rel="stylesheet"
