+++ src/popup/assets/all-caught-up.webp
