+++ src/popup/popup.html
@@ -20,7 +20,7 @@
 			<header>
 				<div class="logo">
 					<img
-						src="../../assets/icons/icon128.png"
+						src="../assets/icons/icon128.png"
 						alt="Toppings Logo"
 						draggable="false"
 					/>
